aider: Update `dump_samples.py` to store `expected_output` as a YAML multi-line string in `temp.yaml`.
# Aider chat conversation:

USER: the temp.yaml file should have the expected_output value as a yaml multi-line string.

ASSISTANT: To modify the `dump_samples.py` script so that the `expected_output` value is a YAML multi-line string, here is the diff:

```diff
--- tests/dump_samples.py
+++ tests/dump_samples.py
@@ -8,11 +8,11 @@
     return template.render(transaction=transaction_data).strip()

 with open('transaction_samples.yaml', 'r') as file:
     samples = yaml.safe_load(file)

-for sample in samples['transaction_samples']:
-    sample['expected_output'] = """{}""".format(render_transaction(sample['transaction_data']))
+for sample in samples['transaction_samples']:
+    sample['expected_output'] = render_transaction(sample['transaction_data'])

 with open('temp.yaml', 'w') as file:
-    yaml.dump(samples, file)
+    yaml.dump(samples, file, default_flow_style=False, allow_unicode=True)
```

This change will ensure that the `expected_output` is stored as a multi-line string in the `temp.yaml` file. The `yaml.dump` function is modified to use `default_flow_style=False` to enable multi-line strings and `allow_unicode=True` to allow Unicode characters.